cmd/compile: unwrap OpOffPtr during DSE
DO NOT SUBMIT

This is a quick hack to illustrate my comments in #25132.

There are all kinds of things wrong with it:
Fake XPoses, missing overflow checks, poor names, etc.

I do not intend to improve it; I will leave it to Daniel Marti
or someone else to think about how to do this properly.

Change-Id: I9cc610be3cf1f6dc83b5238051602e6b777e15c4