Update patch set 2
Patch Set 2: Code-Review+2

LGTM. 

It took me a while to figure out why the old code fails. It seems asmoutnacl updates p.Pc for alignment, but we passed op to checkpool, and op.Pc does not include the alignment. I guess if we pass p.Pc-op.Pc+m-4 as the size arg to old checkpool, it would work? (Not saying we should do it this way -- the new code is better.)

Yeah, the old expression in checkpool is really hard to read. I can guess most of them, but not really for sure.

Patch-set: 2
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2