runtime: convert exit to use pthread library
Now we no longer need to mess with TLS on Darwin 386/amd64, we always
rely on the pthread library to set it up. We now just use one entry
in the TLS for the G.
Return from mstart to let the pthread library clean up the OS thread.

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: Iccf58049d545515d9b1d090b161f420e40ffd244