Update patch set 1
Patch Set 1:

This is split out from CL 43157, which will have to wait for Go 1.12.


Note that in that CL, Philip Hofer wrote of these rules:

This might be an anti-optimization on arm64 and mips, which both have some notion of branch-if(-not)-zero. It doesn't appear that they undo this transformation in the arch-specific rewrite rules either. Probably worth filing an issue so that we don't drop the ball there.


And you replied:

I think they do the right thing given this transformation. The rules (arm64 example)

(Eq64 x y) -> (Equal (CMP x y))
(If (Equal cc) yes no) -> (EQ cc yes no)

Should just result in a CMP/BEQ pair.  That's not any worse than a SUB/branch-if-zero, is it?
If the SUB was used elsewhere, maybe.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>