encoding/json: include property name for marshal errors
When a marshal error is encountered, the error message (and underlying `MarshalerError`) does not include any context about where the error was encountered, even if the value is complex like a map, slice or struct.

This change will include that metadata in both places, allowing a caller to have more context in the event of a JSON marshal error.

Fixes #25153

Change-Id: Iddc737a0659e6d81e2689a09856eb4cc663afc57
GitHub-Last-Rev: 084749ef3d2fe7409366caa22a5c81e12dd44cc2
GitHub-Pull-Request: golang/go#25154