cmd/compile: plumb prologueEnd into Dwarf
DO NOT SUBMIT, something is wrong with line numbers.

If the line number glitch can be fixed in the next day or
two, this would be good to have in 1.11 for Delve.

This steals 2 more bits from lico, reducing the maximum
line number to about 1 million.  Positions can also be
tagged as "end of prologue" and "beginning of epilogue",
though this fortunately happens after optimization.

The dwarf attribute is attached to the proper PC, but at
some point the line numbering gets slightly off (sometimes,
negative).  How/why is still TBD.

Change-Id: I3bda5f46a0ed232d137ad48f65a14835c742c506