Update patch set 4
Patch Set 4:

*sigh* Somehow enabling inlining of functions that call throw is actually slowing things down in David's experiments, so we're not going to do that for now. I don't know why, because the list of functions that get inlined when you do that seems utterly reasonable. But for now we can put this change in.

When it's not crunch time, I think I'll prep a CL that de-hand-optimizes and de-obfuscates a bunch of runtime stuff, and we can try some inlining experiments again. It should be possible for the compiler to do better than the hand-optimized stuff, but perhaps not when the hand optimization is getting in its way.

Patch-set: 4
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>