Update patch set 25
Patch Set 25:

(6 comments)

I wonder whether this could be done as a separate pass, instead of interleaved with the register allocator? I find it hard to understand. This only needs to go through the values and mark certain values on stack, right? Then the register allocator would simply need to skip those values? There need to be a way to communicate this to the register allocator and the code generator, maybe adding a field to Value, or some other way.

Patch-set: 25
CC: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>