Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> * This code is not symmetrical, and people don't read the docs, so they will use it incorrectly.

But the asymmetry is only present if the input slices have different lengths, where the code in master just short-circuits; so this code is strictly an improvement?

> * The compiler is free to optimize away k, and I wouldn't be surprised if it did so today already.

That is why I said here that the code still needs to be rewritten in assembly.

> * The length of the secret is not considered secret itself. In fact, in many applications (eg. HMAC), the length of the secret is prescribed by the standard.

But it could help in the linked issue, where the secret is a variable length passphrase.

Patch-set: 2
Reviewer: Gerrit User 15048 <15048@62eb7196-b449-3ce5-99f1-c037f21e1705>