Update patch set 23
Patch Set 23:

> Patch Set 23:
> 
> (27 comments)
> 
> There are several implementations of the "move Stmt to the next viable Value" pattern in here, which makes me wonder two things:
> 
> 1. Can we abstract this? It seems like a fair amount of effort each time, and there are subtle differences between some of the implementations and I don't know if those are intentional.
> 
> 2. They're all based on the order of Values within a Block, but I thought that was generally meaningless until we did layout.

For purposes of adding to the permanent record (since we discussed this earlier today), the values are ordered-ish.  They start ordered, and they become ordered at scheduling passes.
Optimization passes that creation new values tend to do so in order, so what you get is several repeats of the ordering, roughly.  If you can find a plausible new home for a statement boundary nearby, that's good enough.

The one screw case that I am worried about is a single-block loop in which the values for the loop header appear at both beginning and end of the loop; I'd like to preserve both of those.  Arguably, I should add a second kind of statement boundary (we have the bits) for "IsSecondStmt" and that just might do it.  This won't generalize to unrolled loops, and properly unrolled loops could potentially have the entire loop body replicated in a single block.

However, I think you're right, and I should certainly try the experiment and see if there are any notable failures and/or improvements.

There are also two variants on the statement boundary placement that don't depend on orders in a block; when doing CSE, there are two related values to a value being removed (its replacement, and the value for which it is an arg), and deadcode, which constructs an ordering of the values which can be used to ensure that the "first" value in a statement gets the boundary marker.

I had a thought just now, which was that concerns about the exact value chosen for "start" could be deferred until scheduling.  There are short-term (1.11) and long-term versions (1.12+) versions of this.  Suppose that blocks are associated with the statements beginning within them, and the block optimizations (in particular, fuse and deadcode) keep track of block names/contents.  Then the scheduling pass, for each block, will check the lines of values scheduled, and the first one that is in the statement-start table for that block will get the boundary.  Short-term, the values are carried in the block, and boundaries can be noted as they go into the schedule.

Patch-set: 23
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>