Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-2
> > 
> > * This code is not symmetrical, and people don't read the docs, so they will use it incorrectly.
> 
> But the asymmetry is only present if the input slices have different lengths, where the code in master just short-circuits; so this code is strictly an improvement?
> 
> > * The compiler is free to optimize away k, and I wouldn't be surprised if it did so today already.
> 
> That is why I said here that the code still needs to be rewritten in assembly.
> 
> > * The length of the secret is not considered secret itself. In fact, in many applications (eg. HMAC), the length of the secret is prescribed by the standard.
> 
> But it could help in the linked issue, where the secret is a variable length passphrase.

Consider a random 8 byte password. That has roughly 56 bits of entropy (there is 2^64 possibilities but a somewhat less because not all characters can be in a password.)

If you discover the length of the password, that gives you just 3 bits of information, which doesn't appreciably help discover the contents of the password. 

It's much more important to protect the contents of the password, because you can use a timing  measurements to discover the secret byte by byte.

Patch-set: 2
Reviewer: Gerrit User 5893 <5893@62eb7196-b449-3ce5-99f1-c037f21e1705>