Update patch set 6
Patch Set 6:

(6 comments)

I've uploaded some changes which address some of the comments. On my machine, the parallel benchmark is three times as fast with this implementation compared to the existing implementation: 51.8 ns/op compared to 148 ns/op. At the moment, there's a performance hit for the serialized approach, but there are some pretty low-hanging fruit here. I opted for a simple implementation as a first pass before optimizing it, and I'm confident we could bring that back down.


I agree that it would be fruitful to work on an improved RNG for Go 2, but I'd also like to get a version of math/rand that performs better under parallel execution before that. The global math/rand functions are currently an important bottleneck for our production code, and from the issues linked, it appears others are in similar positions as well.

Patch-set: 6
Reviewer: Gerrit User 22735 <22735@62eb7196-b449-3ce5-99f1-c037f21e1705>