Update patch set 6
Patch Set 6: Code-Review-1

This still has impacts on sequential rand calls in some circumstances that are not predictable by end users. For scientific code users, this can have significant impacts for repeatability.

The issue here is that there is globally mutated state that alters the behaviour of the PRNG but which clients of the package cannot alter; if a package uses rand functions in parallel calls and then the final client re-seeds with rand.Seed, the behaviour is still in the parallel call state and cannot be returned to the original sequential state.

Is there a reason your client code can't just manage multiple *rand.Rand values?

Patch-set: 6
Reviewer: Gerrit User 6480 <6480@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1