Update patch set 1
Patch Set 1:

(1 comment)

The reason for being stricter with the symbols that are processed in defdwsymb is that, before this change, defdwsymb relied on the symtab step to mark a lot of symbols as NotInSymbolTable. This has also the nice side effect of removing statictmp and $f64/$f32 variables from debug_info.
The only negative side effect, AFAIK, is that handwritten assembly can (and does) place symbols directly into .rodata, for example the exprodata and masks symbols used by the math package, and this CL removes them from debug_info.

Given that I'm late for 1.11, it's a fairly invasive CL and the bug it's fixing is not very important I think it can wait for 1.12.

Patch-set: 1
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>