Update patch set 41
Patch Set 41:

This CL as is is unreviewable even if I think it properly implements
Kim&Kutzners algorithm.

If the proposal committee agrees that a better algorithm based on
"Ratio Based Stable In-Place Merging" by Kim and Kutzner is needed
I think you will have to restart in the way Robert described:
In an incremental way where each increment by itself is reviewable
and correct.

The K&K merging is a 4 step process:
 1. Block distribution storage extraction (BD)
 2. buffer extraction for H&L (BUF)
 3. Block rearangement
 4. Local merges
 5. Final sweep up.
Which can be broken down into natural increments:

Lets assume temporary that a) we use SymMerge for step 4
and b) won't do neither step 1 nor step 2.
So we focus und block rearangement (BR) and local merge (LM) _only_.
What does BR need? It needs just one piece of information:
"Which is the current minimum block?"
Same for LM: All it needs to know is "Wich block are w-blocks?".

My suggestion is:
Make an object (lets call it housekeeping) which can answer these
two questions: "Minimum block?" and "Wich are w-blocks?" through
two methods. The inner workings are opaque.
To answer these question it needs to know what happens during BR
so it will need methods through wich it can be informed about w-block
placements during BR.

Let's keep houskeeping simple and correct: Instead of a block
distribution storage (BS) to answer "which are w-blocks?" let's
use a dynamically allocated []bool. For finding "minimal block?"
use what ever is appropriate, e.g. a dynamically allocated []int.

Encapsulate BR into a testable function. Write tests for BR.
Write tests for BR which test the 3 properties required by the
paper for BR.

This is a very general rule: Break the code down in small, isolated
function and write tests for these function. This makes it clear
what the functions should do, how the work and will give confidence
that they are okay. All this helps understanding the code in general,
during review and during maintenance of the code.

With such an infrastructure in place you can implement BR and
LM|SymMerge and the implementation
will be straightforward, modular and thus testable and reviewable.

Next iteration:
Implement Hwang & Ling (HL) with an _explicit_ buffer argument
(callers of HL must provide this buffer).
Add tests for HL. Lot's of tests. With a standalone function
you can come up with small test cases and make sure test coverage
is high.

Next iteration:
Implement BUF extraction in step 1 and add tests for this buf
extraction. Store information about this merge buffer un houskeeping
(no manual indices juggling). Housekeeping knows everything about buf:
Is is big enough? Where is it? Etc.
Write tests for BUF extraction. Lot's of them.
Use HL during LM whenever possible (buf bug enough, ...)
Implement the necessary sweepup of step 5.

Such a change would be reviewable: BUF by itself is a small function
backed by test. HL is tested already. Sweep up is simple.

Next increment:
Use a fixed size BD storage or an extracted one instead of a []bool.
Again this is a small, reviewable change. It will add a BD extraction
step with appropriate tests and it's effect on the rest of the
algorithm is 0 as it is encapsulated in housekeeping.

Next iteration:
Do the optimizations mention in the paper.

Next iteration:
If (and only if) the inlining performed by the compiler is
not good enough: Do manual inlining.



I do not think that code review should happen out-of-band through
a github repo. With an incremental approach as proposed by Robert and
described above each CL can be made small and reviewable.
You could try to keep the number of patchsets small by addressing
all review concerns together (and marking them Done) before uploading
a new patchset.  

So let's wait for the outcome of the proposal committee.

Patch-set: 41