Update patch set 41
Patch Set 41:

Some quick responses:

> If the proposal committee agrees that a better algorithm

Surely you mean better (more modular and less efficient) *code* with the algorithm staying the same?!?

> In an incremental way where each increment by itself is reviewable
> and correct.

Not saying I think this is a bad approach, but note that this means much more work for me *and* more work in total for whoever reviews it... But I guess that must be OK, since you are the only one willing to review.

> The K&K merging is a 4 step process:
>  1. Block distribution storage extraction (BD)
>  2. buffer extraction for H&L (BUF)
>  3. Block rearangement
>  4. Local merges
>  5. Final sweep up.

Actually, it is natural in a merge *sort* to only do steps 1, 2, and 5 once in merge sort level (as documented in comments at the beginning of func stable); thus func merge (which implements K&K merging proper) only implements steps 4 and 5. This does not conflict your plan below, but I thought necessary to note it for clarity.

> Make an object (lets call it housekeeping) which can answer these
> two questions: "Minimum block?" and "Wich are w-blocks?" through
> two methods. The inner workings are opaque.

It also would need to expose about 6 integers because we need to know where are the bounds of BDS and MIB when we are merging the merge sort blocks (so we would not try to merge over BDS or MIB)

> To answer these question it needs to know what happens during BR
> so it will need methods through wich it can be informed about w-block
> placements during BR.

I am sure everybody is aware, but this is going to be an efficiency loss. Considering that we have to expose the integers anyway like I just said, I do not see what is wrong with the current way this is done, that is just calling rotate and de/incrementing some integers.

> Encapsulate BR into a testable function. Write tests for BR.
> Write tests for BR which test the 3 properties required by the
> paper for BR.

You may be too keen on testing considering TestStableBM already has total coverage (when the relevant line is uncommented in testBentleyMcIlroy). There is probably some code paths not tested, but block redistribution is not among them, BR being wrong would fail TestStableBM.

> Next iteration:
> Implement Hwang & Ling (HL) with an _explicit_ buffer argument
> (callers of HL must provide this buffer).

Why the _explicit_ buffer argument? (Also, this is called simpleMergeBufBigSmall in my code and does not use the H&L merge, which is OK in the context of the merge sort.)

> Add tests for HL. Lot's of tests. With a standalone function
> you can come up with small test cases and make sure test coverage
> is high.

Umm, again, test coverage is already total, and a faulty merge would fail tests.

> Next iteration:

Are you really bent on reviewing separate BDS and MIB CLs? Because that seems a bit painful to implement.

> Use a fixed size BD storage or an extracted one instead of a []bool.
> Again this is a small, reviewable change. It will add a BD extraction

Nitpick - BDS is just "assigned" (we note where it is), not extracted.

> step with appropriate tests and it's effect on the rest of the
> algorithm is 0 as it is encapsulated in housekeeping.

Nitpick: except for the final sweep up.

> Next iteration:
> Do the optimizations mention in the paper.

The "floating hole" optimization is not possible given the sort API.

> So let's wait for the outcome of the proposal committee.

OK.

Patch-set: 41
Reviewer: Gerrit User 15048 <15048@62eb7196-b449-3ce5-99f1-c037f21e1705>