Update patch set 6
Patch Set 6:

Could you elaborate on the circumstances under which this would have impact on purely sequential uses of math/rand? My belief is that this approach should preserve the existing behavior of math/rand while still providing the lock-free performance for concurrent uses.


> if a package uses rand functions in parallel calls and then the final client re-seeds with rand.Seed, the behavior is still in the parallel call state and cannot be returned to the original sequential state.


I'm curious how much existing code relies in this particular pattern of behavior - a fixed number of parallel calls to the global RNG, followed by a serialized re-seed and purely sequential subsequent calls to the global RNG - which also cannot handle the output of the RNG change.

That said, this is an easy problem for us to fix; the global reseeding operation should switch back to the sequential version of the algorithm.


> Is there a reason your client code can't just manage multiple *rand.Rand values?

Unfortunately, that is the case. As mentioned on #25057 and #24121, the current design of the math/rand library makes it most ergonomic for library developers to use the global math/rand functions. This is a problem for applications which import more than one package which in turn imports math/rand, because it increases the probability of mutex contention exponentially.

As the application developer, there's no way for me to reduce the mutex contention between the libraries I import. Libraries almost never have a way to specify non-default threadsafe RNGs that should be used in place of the default, because there's no constructor in the standard library for generating a non-default, threadsafe RNG, nor is there a common interface type definition for them all to standardize around.



This is a real problem for performance-sensitive applications at scale, and given the current design of the standard library, the only way for those applications to manage the mutex contention that they experience is to address the mutex contention within the standard library itself.

Patch-set: 6
Reviewer: Gerrit User 22735 <22735@62eb7196-b449-3ce5-99f1-c037f21e1705>