Update patch set 2
Patch Set 2:

I was thinking that there is another possible way to fix this bug. Specifically, leave mangling where it is (more or less) and make DWARF symbols for types as soon as we se a `type.` symbol, before deadcode.

The downsides are that:

1. it would be doing more work because some types are discarded by deadcode
2. I'm pretty sure it would expose a latent bug in the code that generates DWARF types. I've seen it happen while I was doing this version, that code isn't meant to see certain type symbols.
3. there would have to be some code after deadcode that actually decides which DWARF types go inside dwtypes and into debug_info.

The upside is that it would be closer to producing DWARF in the compiler and it wouldn't need this scary late symbol Rename.

Which is better?

Patch-set: 2
Reviewer: Gerrit User 5821 <5821@62eb7196-b449-3ce5-99f1-c037f21e1705>