Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Could you elaborate on the circumstances under which this would have impact on purely sequential uses of math/rand? My belief is that this approach should preserve the existing behavior of math/rand while still providing the lock-free performance for concurrent uses.

The code here would exercise this: https://play.golang.org/p/wGECHZkJfkM

```
    package main

    import (
    	"fmt"
    	"math/rand"
    	"sync"
    )
 
    // otherpackage
    func otherpackage_middledot_expensiveSetupFunc(i int) {}

    func /*otherpackage_middledot_*/ init() {
    	var wg sync.WaitGroup
    	for i := 0; i < 10; i++ {
    		wg.Add(1)
    		go func() {
    			otherpackage_middledot_expensiveSetupFunc(rand.Int())
    			wg.Done()
    		}()
    	}
    	wg.Wait()
    }

    // userpackage

    func main() {
    	rand.Seed(1)
    	for i := 0; i < 10; i++ {
    		fmt.Println(rand.Int())
    	}
    }
```

I can imagine a fairly significant set of scientific simulation code that would follow this pattern. In these cases having a simulation that is completely repeatable is critical for reproducible science.

> I'm curious how much existing code relies in this particular pattern of behavior - a fixed number of parallel calls to the global RNG, followed by a serialized re-seed and purely sequential subsequent calls to the global RNG - which also cannot handle the output of the RNG change.

The thing is that you cannot know that some package in your dependencies has not done what the code above does, performing concurrent rand calls during set up.

> That said, this is an easy problem for us to fix; the global reseeding operation should switch back to the sequential version of the algorithm.

If this can be done safely then it would be OK I guess, though the design here is pretty baroque.

> 
> > Is there a reason your client code can't just manage multiple *rand.Rand values?
> 
> Unfortunately, that is the case. As mentioned on #25057 and #24121, the current design of the math/rand library makes it most ergonomic for library developers to use the global math/rand functions. This is a problem for applications which import more than one package which in turn imports math/rand, because it increases the probability of mutex contention exponentially.
> 
> As the application developer, there's no way for me to reduce the mutex contention between the libraries I import. Libraries almost never have a way to specify non-default threadsafe RNGs that should be used in place of the default, because there's no constructor in the standard library for generating a non-default, threadsafe RNG, nor is there a common interface type definition for them all to standardize around.
> 
> 
> 
> This is a real problem for performance-sensitive applications at scale, and given the current design of the standard library, the only way for those applications to manage the mutex contention that they experience is to address the mutex contention within the standard library itself.

The issue here is not really the design of the standard library's math/rand package, but rather that other people insist on using the global rand.Rand for their PRNG work in libraries. They shouldn't. When a library provides a way to hand in a *rand.Rand or a rand.Source then the issue of contention can be reduced to 0 for a well designed client package.

Note that there is a race in the code here as mentioned by bradfitz - confirmed by running the code above with your change here.

Patch-set: 6
Reviewer: Gerrit User 6480 <6480@62eb7196-b449-3ce5-99f1-c037f21e1705>