Update patch set 15
Patch Set 15:

Is there something else that needs to be done for this review to move forward?
Keith suggested rewriting some the rules for the purpose of making the rewritePPC64.go file smaller, but I spent many hours trying to do that without success (i.e., some things were no longer matched), and I didn't see how his suggestion applied to ppc64le. The size of this file is reasonable with respect with other platforms. If the size of the rewrite file is a concern then the rulegen.go should be changed to avoid generating all combinations when there are more than 3 or 4 operands if it doesn't make sense to match all orders (regardless of the top level op). That would reduce the size of all rewrite files. Seems like you could easily add a condition to indicate not to reorder the operands. Whether it can be commuted shouldn't just depend on the op but also on the operands because some orders won't ever occur.

Patch-set: 15
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>