cmd/internal/obj/wasm: avoid invalid offsets for Load/Store
Offsets for Load and Store instructions have type i32. Bad index
expression offsets can cause an offset to be larger than MaxUint32,
which is not allowed. One example for this is the test test/index0.go.

Generate valid code by avoiding the invalid offset and emit an
"Unreachable" instruction so the code is guaranteed to fail at
runtime.

Change-Id: Ie90adcbf3ae3861c26680eb81790f28692913ccf