Update patch set 28
Patch Set 28:

(7 comments)

> Patch Set 25:
> 
> (6 comments)
> 
> I wonder whether this could be done as a separate pass, instead of interleaved with the register allocator? I find it hard to understand. This only needs to go through the values and mark certain values on stack, right? Then the register allocator would simply need to skip those values? There need to be a way to communicate this to the register allocator and the code generator, maybe adding a field to Value, or some other way.

Have you looked at the possibility of moving it out of the register allocator? Thanks.

(Since this CL is an optimization, not a prerequisite of other wasm CLs, maybe move this CL to the top of the CL stack, so it is clear that other CLs can be submitted without this one in first.)

Patch-set: 28