Update patch set 1
Patch Set 1:

There's a concern that existing code expects a String rather than a RawString due to the existence of this bug; and fixing this might break such code. It's always possible to detect if we have a RawString or a regular string by looking at the first character of the token value.

I guess we will find out.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>