Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review+2
> > > 
> > > LGTM. Maybe explain in the commit message that this is split out of the previous CL to explain why it is going in after the freeze?
> > 
> > I'm actually not sure that it is ok to submit this after freeze. Keith?
> 
> 1.11 is now open only for bug fixes and CLs that were outstanding as of the freeze.
> Performance-only stuff will have to wait for 1.12.

There’s one argument (perhaps weak) for doing this in 1.11, which is that amd64 has more code miles put on it than s390x, which is currently the only arch that has this optimization. Doing this on amd64 too thus increases the likelihood of catching any bugs with this optimization before the release.

Patch-set: 1
CC: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>