Update patch set 1
Patch Set 1: Code-Review-1

Thank you for this change. The development tree is currently frozen ahead of the 1.11 release [0]. To proceed with this change please consider the following suggestions

a. if this change is response to an issue targeted at Go 1.11, please include the text "fixes #NNN" where NNN is the issue this change relates to [1].
b. if this is a performance improvement, please include benchmark results using benchcmp [2] showing the potential improvements this change makes. Please be aware that this change may not be accepted during the Go 1.11 code freeze.

0: https://github.com/golang/go/wiki/Go-Release-Cycle
1. https://golang.org/doc/contribute.html#commit_changes
2. https://godoc.org/golang.org/x/tools/cmd/benchcmp

Patch-set: 1
Reviewer: Gerrit User 5150 <5150@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1