Update patch set 32
Patch Set 32:

(3 comments)

> Patch Set 31:
> 
> How do we want to continue on this CL?

Sorry for the delay. I'm traveling and in a meeting.

> Which code exactly would be moved out? Only the part that sets the OnWasmStack flag? I think we wouldn't get that much by moving those lines and they are quite related to register allocation...

My most concern is the change in allocValToReg. I think it is not very easy to follow, given the inherent complexity of the register allocator. I wonder if we could check OnWasmStack and skip values on a higher level? Maybe that is not easy. I'm still looking.

Patch-set: 32